# backend/docs/social-api-draft.openapi.yaml
...
openapi: 3.0.3
info:
  title: STROITECH Social API (draft)
  version: "0.1.0"
  description: >
    Планируемые маршруты. Префикс API в проде обычно /api (как у текущих /api/users, /api/feed).
servers:
  - url: /api
...
paths:
...
  /communities/{id}/albums:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [communities, albums]
      summary: Альбомы, привязанные к сообществу
      responses:
        "200":
          description: OK
...
  /users/{id}/albums:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [albums]
      summary: Список альбомов пользователя
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
  /albums:
    post:
      tags: [albums]
      summary: Создать альбом (владелец = JWT)
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: OK
  /albums/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [albums]
      summary: Альбом с фотографиями
      responses:
        "200":
          description: OK
    put:
      tags: [albums]
      summary: Обновить альбом (только владелец)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    delete:
      tags: [albums]
      summary: Удалить альбом (только владелец)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /albums/{id}/photos:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [albums]
      summary: Добавить фото (URL после POST /upload/album)
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: OK
  /albums/{id}/photos/{photoId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: photoId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [albums]
      summary: Удалить фото из альбома
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /albums/{id}/photos/reorder:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      tags: [albums]
      summary: Установить порядок фото (массив id)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
  /upload/album:
    post:
      tags: [albums]
      summary: Загрузка изображения в префикс albums/{userId}/
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }